Remove modules and routines that have to do with OpenDA/OpenMI
Remove more code that is connected to OpenDA/OpenMI. I have not removed any files yet, as I was not quite sure what the consequences are, but we can certainly remove more.

Note: I also came across an interface to MT3D. Is that used at all? If not, then that can go as well.

Is there still a role for delwaq_lib? It seems geared to the interfacing with OpenDA. But again, I am not entirely sure.