Correct the selection of the dispersion coefficient
In the routine dlwq19, the wrong dispersion coefficient was selected for unstructured grids where noq2 is equal to zero. This led to incorrect results and to a very significant drop in performance. The change takes care of this. I could have done it in a different way, but the comments should make the logic behind the change clear.