UNST-7889 removed erroneous call to updateBalance from finalize_single_timestep
A call to updateBalance (which in turn calls reduce_bal to reduce the water balance integrated values across the processes) was erroneously introduced into flow_finalize_single_timestep. As this already happens in flow_finalize_user_timestep, the reduction now happened twice, effectively multiplying the water balance by the number of processes. After removing this call, the epic branch now gives the same results as the mian (and the reference) in test case e02_dflowfm\f114_parallel_1D\c040_korte-woerden-1d.

Note: the exception in reduce_statistical_output where output items defined on UNC_LOC_GLOBAL are skipped is still there. It doesn't appear to be necessary, but that's likely coincidence because it is called after reduce_bal instead of before (which I think is arbitrary). Right now it helps nor harms, but if the order of those calls is ever reversed, then it will be necessary to have this exception.