UNST-6079 fix bug in prop_get_set_integers
Fixed a bug in prop_get_set_integers that caused the reported "length of moving average window affects 'current' output" bug.

With the new statistical output, MDU output parameters such as wrihis_velocity, which used to be simply 0 or 1, can also be strings, e.g. "current, average, max(5)". However, while statistical output's own MDU reading function (m_output_config/scan_input_tree) can handle this, the old function (properties/prop_get_set_integers) cannot, as it expects only an integer. If the parameter did not contain any integer (e.g. "current, average, max"), then it would stick to the default value, which was 1, so all was well. However, in the case of "current, average, max(5)", it would read the 5 as the integer, and thereby set wrihis_velocity equal to 5. This messed up the check in fill_valobs: IF (wrihis_velocity == 1) THEN [calculate velocity magnitudes]. As it happens, the velocity magnitudes are also (re)calculated in some other parts of the model, but only once every user time step. Therefore, the input for the statistical output calculation would be wrong, but not empty, causing the weird bug.

All this is now fixed by allowing prop_get_set-integer to optionally allow statistical output keywords as input instead of integers. Any string containing 'current', 'average', 'min', or 'max' is then translated to an integer value of 1, so that the old wrihis parameters are still valid.