Deprecate the run_dflowfm_processes scripts
There is no reason for having scripts that set the options for the online water quality options and scripts that do not explicitly set them. As a first step, that functionality is incorporated in the standard script and the scripts run_dflowfm_processes simply call that standard script.

Note: if you do not use the online water quality processes, these options are not even reported (excpet as part of the entire command-line)