Correct the order of the arguments
The order of the arguments to extract_waq_attribute as they appeared in the section to write warnings about insufficient data was wrong. This would have led to perhaps a bit more messages than necessary (the active cells were not identified properly) but it might have been responsible for divisions by zero.