UNST-8957: NetCDF variable time: check dimensions
ec_netcdf_timeseries.f90: Variable "time" has to be found when reading the NetCDF variables. Checking "standard_name=time" is not enough because multiple variables might have this. Added: check that the number of dimensions is 1 and that the dimensionName is also "time".
Function "ncu_inq_dimname" is added, calling the existing "nf_inq_dimname", although this function was not added yet to the interface.