UNST-9154: Fix partitioning of net-files with bed level at faces in context of dry points/areas (#255)
* UNST-9154: Partitioning a model with bed levels specified on the net-file consists of the following steps:
1) read mesh and bed level from net-file
2) remove dry points
3) determine the desired partitioning of the mesh
4) for each partition:
4a) clip the network (nodes and edges) to the partition
4b) reconstruct the faces from the network geometry
4c) remove dry points
4d) determine the bed levels at faces for the partition based on iglobal_s indexing from the mesh after step 4c to the mesh after step 2.
For this to work we need to update blcell array in step 2, but not in the equivalent step 4c. Hence we need a flag "update_blcell" that needs to be passed down through a couple of calls. For lack of information, we set update_blcell equal to .false. (old behaviour) in all places that are unrelated to our current case.

* UNST-9154: add a test cases e02/f014_/c102 and c103 to verify that partitioning of net-files with bed level specified at faces works correctly together with the clipping of dry points/areas. Repeat of commits on Sep 6/8, 2025.

* add description for new update_blcell argument
check on non-empty allocation of blcell before changing it

* add missing definition of "dimr_platform_dependent" test cases: normal dimr run with platform dependent results ... needed for METIS partitioning

* replace "dimr_platform_specific" testCase configuration by a "generic_platform_specific" (and a "generic") configuration since the partitioning tests are not running dimr but dflowfm.
comment added to explain why we need to have a test case with platform specific results (... it's due to METIS)