Waq/bugfix/delwaq 622 stable numbering boundaries (#374)
* First step to implement partition-independent numbering

As a first step, sorting and renumbering independent of the partitioning. It does not work completely, as yet, but it is a start.

* Correct use of the various indices

The waqmerge program (or rather its underlying modules) use quite a few indices and it was not straightforward to sort these out. This commit leads to the right and stable numbering.
(Note: still a few debug write statements)

* DELWAQ-622: Remove debug output

* Count the number of boundary cells

Count the number of boundary cells per layer and correct a small mistake.