Waq/bugfix/delwaq 356 fix fluff model for metals (#266)
* Add adsorption to POC and subsequent sedimentation to S2

The "fluff" model allows you to specify that part of the particulate matter ends up in the S2 layer directly. This was correctly extended to metals adsorbed to inorganic matter, but not to organic micropollutants or to adsorption onto organic material. This rectifies that.

* Add the missing flux parameters for the buffer model

Because the processes dealing with the adsoroption onto POC require a change to the cofiguration, we need to add them to the configuration.

* DELWAQ-356: Set a default value for the new fluxes to S2

For the process SUM_SEDIM we need several new fluxes - fSedPOCxS2. These did not have defaults, so many test cases were failing. This commit repairs that: the fluxes are set to zero by default.

* DELWAQ-536: Correct configuration for viruses

The processes "Sed_OMP" and "Sed_Virus1/2/3" were missing a few entries.

* Solve the issue with at least one test case

In the routine SOMSED the arrays of pointers into the PMSA array was re-initialised, which destroyed the preparation for the loop over the vertical exchanges.

* Correct the pointering for segment-wise parameters

* Update the reference files for PROTIST test cases

Required because of the bugifx.

* Update the reference files for four remaining tests

* Update items.csv

* Friendlier error messages

* Friendlier error messages

* conflict solved.

* solve the conflict

* Revert "solve the conflict"

This reverts commit 6ce96874a351d848a848b573a5a039f3d359db50.

* update references.

* DELWAQ-356: Merging changes to the watage routine

I have no idea how things have crossed each other, but the simplest way to solve the failures to the unrelated testcases is to update the WATAGE process routine,.

* Update of the proc_def files

---------

Co-authored-by: Qinghua_ye <qinghua.ye@deltares.nl>
Co-authored-by: Qinghua Ye <113536652+ye4deltares@users.noreply.github.com>